requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
bokeh==3.3.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.2
certifi==2023.7.22
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorcet==3.0.1
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cubinlinker-cu11==0.3.0.post1
cucim==23.10.0
cuda-python==11.8.2
cudf-cu11==23.10.0
cugraph-cu11==23.10.0
cuml-cu11==23.10.0
cuproj-cu11==23.10.0
cupy-cuda11x==12.2.0
cuspatial-cu11==23.10.0
cuxfilter-cu11==23.10.0
Cython==3.0.4
dask==2023.9.2
dask-cuda==23.10.0
dask-cudf-cu11==23.10.0
datashader==0.16.0
defusedxml==0.7.1
distributed==2023.9.2
exceptiongroup==1.1.3
fastjsonschema==2.18.1
fastrlock==0.8.2
filelock @ file:///croot/filelock_1672387128942/work
fiona==1.9.5
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
geopandas==0.14.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
h5py @ file:///croot/h5py_1691589708553/work
hmmlearn @ file:///home/conda/feedstock_root/build_artifacts/hmmlearn_1696709150716/work
holoviews==1.18.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.8.0
isoduration==20.11.0
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib @ file:///croot/joblib_1685113087166/work
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_proxy==4.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
lazy_loader==0.3
linkify-it-py==2.0.2
llvmlite==0.40.1
locket==1.0.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.7
multidict==6.0.4
multipledispatch==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
networkx @ file:///croot/networkx_1690561992265/work
numba==0.57.1
numpy==1.24.3
nvtx==0.2.8
opencv-python==4.7.0.72
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
panel==1.3.0
param==2.0.0
partd==1.4.1
Pillow==9.5.0
platformdirs==3.11.0
prometheus-client==0.17.1
protobuf==4.24.4
psutil==5.9.6
ptxcompiler-cu11==0.7.0.post1
ptyprocess==0.7.0
pyarrow==12.0.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyct==0.5.0
Pygments==2.16.1
pylibcugraph-cu11==23.10.0
pylibraft-cu11==23.10.0
pynvml==11.4.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyproj==3.6.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pyviz_comms==3.0.0
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==25.1.1
raft-dask-cu11==23.10.0
referencing==0.30.2
requests @ file:///croot/requests_1690400202158/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rmm-cu11==23.10.0
rpds-py==0.10.6
scikit-learn @ file:///croot/scikit-learn_1690978916802/work
scipy @ file:///croot/scipy_1696543286448/work/dist/scipy-1.11.3-cp310-cp310-linux_x86_64.whl#sha256=16a8c87e543daeb96d1310b2283f542fef2de84ee7090f60187eb71f71cb430c
Send2Trash==1.8.2
shapely==2.0.2
simpervisor==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sympy @ file:///croot/sympy_1668202399572/work
tblib==3.0.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.12.0
treelite==3.9.1
treelite-runtime==3.9.1
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
uc-micro-py==1.0.2
ucx-py-cu11==0.34.0
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1698257533958/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
xarray==2023.10.1
xyzservices==2023.10.1
yarl==1.9.2
zict==3.0.0
zipp==3.17.0
